+++ CHANGELOG.md
@@ -4,6 +4,15 @@ CHANGELOG
 AWS-TOWER
 -----
 
+4.4.1
+-----
+
+2023/01/10
+
+### New feature
+  - Add Lambda object, displayed in draw for API Gateway
+  - Add `-o|--output` to save JSON audit and discovery in the specified file
+
 4.4.0
 -----